Refactor to new default supportfile names
