+++ test/cypress/support/e2e.ts
@@ -1,0 +1,18 @@
+import './commands'
+import './setup'
+
+declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace Cypress {
+    interface Chainable {
+      expectChallengeSolved: (value: { challenge: string }) => void
+      login: (value: {
+        email: string
+        password: string
+        totpSecret?: string
+      }) => void
+      eachSeries: any
+
+    }
+  }
+}
